+++ packages/babel-preset/CHANGELOG.md
@@ -9,7 +9,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Created a `common` babel-preset [[#231](https://github.com/Shopify/web-foundation/pull/231)]
+- Created an experimental `common` babel-preset [[#231](https://github.com/Shopify/web-foundation/pull/231)]
 
 ### Changed